worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.2",
    "uvicorn>=0.38.0",
    "httpx>=0.27.0",
    "google-genai>=1.57.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.15.0",
    "qdrant-client>=1.12.0",
    "pymupdf>=1.25.0",
    "beautifulsoup4>=4.13.0",
    "python-multipart>=0.0.20",
]

[tool.poe.tasks]
start = "uvicorn main:app --reload"
openapi = "python scripts/export_openapi.py"
lint = "ruff check ."
format = "ruff format ."
check = "basedpyright ."

[dependency-groups]
dev = [
    "poethepoet>=0.37.0",
    "ruff>=0.9.4",
    "basedpyright>=1.27.1",
]

[tool.ruff]
# Match app project's printWidth
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (import ordering, like eslint-plugin-import)
    "N",    # pep8-naming (like @typescript-eslint/naming-convention)
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501", # Line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["business", "config", "controllers", "routers", "schemas", "services", "tools"]

[tool.basedpyright]
typeCheckingMode = "recommended"
pythonVersion = "3.11"
include = ["."]
exclude = ["**/__pycache__", "**/myenv", "**/.venv", "**/venv"]
venvPath = "."
venv = ".venv"
reportUnusedVariable = "warning"
reportUnusedImport = "warning"
reportMissingTypeStubs = false
reportUnusedCallResult = false
reportExplicitAny = false  # Allow Any type for flexible schemas
reportAny = false  # Allow Any in kwargs and flexible parameters
reportUnknownVariableType = false  # Third-party packages without stubs

